change: api routes to route handler
